Rename BookFlightPostController to CreateBookPostController
