+++ apps/openflight/backend/config/routes/book-flight.yaml
@@ -1,4 +1,4 @@
 book-flight_post:
   path: /book-flight/{id}
-  controller: CodelyTv\Apps\OpenFlight\Backend\Controller\Books\BookFlightPostController
+  controller: CodelyTv\Apps\OpenFlight\Backend\Controller\Books\CreateBookPostController
   methods: [ POST ]
